refactor(structure): move sources and rename namespace
