+++ src/sha512.cpp
@@ -42,7 +42,7 @@
 #include <algorithm>
 #include <cstring>
 #include <cstdio>
-#include "sha512.hpp"
+#include "hmac_cpp/sha512.hpp"
 
 #define SHA2_SHFR(x, n)    (x >> n)
 #define SHA2_ROTR(x, n)   ((x >> n) | (x << ((sizeof(x) << 3) - n)))
